+++ sync-for-payroll/src/sdk/models/operations/listpushoperations.ts
@@ -7,6 +7,9 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class ListPushOperationsRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
@@ -36,6 +39,9 @@ export class ListPushOperationsRequest extends SpeakeasyBase {
 }
 
 export class ListPushOperationsResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -51,9 +57,15 @@ export class ListPushOperationsResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     pushOperations?: shared.PushOperations;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }