+++ sync-for-payroll/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [typescript v1.1.0] sync-for-payroll
 ### Releases
 - [NPM v1.1.0] https://www.npmjs.com/package/@codat/sync-for-payroll/v/1.1.0 - sync-for-payroll
+
+## 2023-10-04 00:05:40
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.93.2 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.2.0] sync-for-payroll
+### Releases
+- [NPM v1.2.0] https://www.npmjs.com/package/@codat/sync-for-payroll/v/1.2.0 - sync-for-payroll